requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
akshare>=1.11.0
openai>=1.0.0
python-dotenv>=1.0.0
